refactored scrolling js
