+++ css/main.css
@@ -22,6 +22,7 @@ html {
 }
 
 body {
+    -webkit-overflow-scrolling: touch;
     font-family: 'Roboto', sans-serif;
     margin: 0;
     font-weight: 200;
