+++ index.html
@@ -273,7 +273,6 @@
 
         </div>
     </div>
-    <script src="/scripts/scrolling.js"></script>
+    <script src="/scripts/application.js"  type="module"></script>
     <script src="/scripts/sidebar.js"></script>
-
 </body>
