+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "smoothscroll-polyfill": "^0.4.4"
+  }
+}
