+++ scripts/application.js
@@ -1,1 +1,6 @@
-scrolling.js
+// import smoothscroll from 'smoothscroll-polyfill';
+
+// smoothscroll.polyfill();
+
+import { initScrollingNavs } from './plugins/scrolling.js';
+initScrollingNavs();
